Découpage de Profil en composant
