+++ frontend/src/pages/styles/profile.module.css
@@ -26,7 +26,7 @@
     align-content: space-around;
 }
 .buttonApplication {
-    width: 35%;
+    /*width: 35%;*/
     font-size: 1em;
     font-style: normal;
     font-weight: 700;
